shopping cart item count

+++ database/migrations/2021_12_26_031159_create_product_table.php
@@ -18,7 +18,7 @@ class CreateProductTable extends Migration
             $table->timestamps();
             $table->string('name');
             $table->text('description');
-            $table->double('price');          
+            $table->integer('price');
             $table->string('image_path');
         });
     }
